Chat

+++ src/ru/fizteh/fivt/students/verytable/chat/MessageUtils.java
@@ -50,7 +50,7 @@ public final class MessageUtils {
     static byte[] toPrimitive(Byte[] bm) {
         byte[] byteMessage = new byte[bm.length];
         for (int i = 0; i < byteMessage.length; ++i) {
-            byteMessage[i] = bm[i];
+            byteMessage[i] = bm[i].byteValue();
         }
         return byteMessage;
     }
